render_requirements.txt
accelerate==1.13.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
av==17.0.1
bcrypt==5.0.0
bitsandbytes==0.49.2
build==1.5.0
certifi==2026.4.22
charset-normalizer==3.4.7
chromadb==1.5.8
click==8.3.3
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
diffusers==0.38.0
durationpy==0.10
filelock==3.25.2
flatbuffers==25.12.19
fonttools==4.62.1
fsspec==2026.2.0
googleapis-common-protos==1.74.0
grpcio==1.80.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.13.0
idna==3.13
importlib_metadata==8.7.1
importlib_resources==7.1.0
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.5.0
kubernetes==35.0.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.9
mdurl==0.1.2
mmh3==5.2.1
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.3
oauthlib==3.3.1
onnxruntime==1.25.1
opencv-python==4.13.0.92
opentelemetry-api==1.41.1
opentelemetry-exporter-otlp-proto-common==1.41.1
opentelemetry-exporter-otlp-proto-grpc==1.41.1
opentelemetry-proto==1.41.1
opentelemetry-sdk==1.41.1
opentelemetry-semantic-conventions==0.62b1
orjson==3.11.8
overrides==7.7.0
packaging==26.2
peft==0.19.1
pillow==12.1.1
protobuf==6.33.6
psutil==7.2.2
pybase64==1.4.3
pydantic==2.13.3
pydantic-settings==2.14.0
pydantic_core==2.46.3
Pygments==2.20.0
pyparsing==3.3.2
PyPika==0.51.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
qwen-vl-utils==0.0.14
referencing==0.37.0
regex==2026.4.4
requests==2.33.1
requests-oauthlib==2.0.0
rich==15.0.0
rpds-py==0.30.0
safetensors==0.8.0rc0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==5.4.1
setuptools==70.2.0
shellingham==1.5.4
six==1.17.0
sympy==1.13.1
tenacity==9.1.4
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.3
transformers==5.7.0
typer==0.25.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.46.0
watchfiles==1.1.1
websocket-client==1.9.0
websockets==16.0
zipp==3.23.1
